refactor: move the anchors logic to an independent file
